+++ config/parser.go
@@ -3,8 +3,6 @@ package config
 import (
 	"encoding/json"
 	"io/ioutil"
-
-	"github.com/gianarb/steady/core"
 )
 
 func Parse(filePath string) (Configuration, error) {
@@ -18,18 +16,4 @@ func Parse(filePath string) (Configuration, error) {
 		return c, err
 	}
 	return c, nil
-}
-
-type Configuration struct {
-	RConf     RConf                     `json:"rconf"`
-	Frontends map[string]*core.Frontend `json:"frontends"`
-}
-
-func (c *Configuration) GetFrontendByName(name string) *core.Frontend {
-	for key, val := range c.Frontends {
-		if key == name {
-			return val
-		}
-	}
-	return nil
 }
